% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_limma.R
\name{difffeatures}
\alias{difffeatures}
\title{Perform ensemble-based limma or normal limma}
\usage{
difffeatures(
  dat,
  pddat,
  responsevarname = NULL,
  responselevels = NULL,
  confoundings = NULL,
  balanceadj = FALSE,
  samplingmethod = "updn",
  nround = 10,
  seed = 2022,
  k = 5,
  threads = 1,
  method = "Cauchy",
  removereg = NULL,
  featuretype = "probe",
  platform = 450,
  ignorestrand = TRUE,
  tssradius = 1500,
  plot = FALSE,
  isbetaval = TRUE,
  absxcutoff = 0,
  pvalcutoff = 0.05,
  pvalcolname = "adj.P.Val",
  titleprefix = NULL,
  titlesize = 15,
  textsize = 13,
  face = "bold",
  labelnum = NULL,
  annotextsize = 4
)
}
\arguments{
\item{dat}{A matrix with the feature values for samples. Each column is one
sample and each row represents one feature. The row names are the feature
names and the column names should be sample IDs.}

\item{pddat}{Meta data frame. The first column should be the sample names
with a column name as "sampleid", and the remaining should be phenotypic
variables.}

\item{responsevarname}{The column name of the response variable in the data
frame provided to \code{pddat}. It is the variable of interest. Default
is NULL, which means the second column of \code{pddat} is the response.
This response variable should be a binary or continuous variable.}

\item{responselevels}{If the response variable is binary, it will be treated
as a factor, and this parameter is needed to define the factor element
level, so it should be a vector defining this level. Only needed if the
response in \code{pddat} is a character vector but not a factor. In this
case, it will be converted to a factor by the function, and this paramter
\code{responselevels} will be used to define the factor element level. Its
default is NULL, meaning that this element level will be set automatically
following the character order of the elements in the response variable.}

\item{confoundings}{The column name of the confounding factors in the data
frame provided to \code{pddat}. Default is NULL, which means there is no
confounding factor in the data.}

\item{balanceadj}{Whether to perform ensemble-based limma or not. When it
is set as TRUE. The function will perform ensemble-based limma. It means
a sampling process will be used on the samples to generate several base
learner datasets, and the sample groups of each base learner set will be
adjusted during the sampling so that they will have the same size. Then,
limma will be used on each base learner to call the differential features,
and their results will finally be ensembled together to get the result for
the whole data. If it were set as FALSE, normal limma will be performed.
Default is FALSE.}

\item{samplingmethod}{When \code{balanceadj} is TRUE, this parameter will
be needed to determine how to sample the data to get the base learner sets
for the ensemeble mode. If it is set as "updn", to make the sample groups
have the same sample size in a base leaner dataset, a large group will be
down-sampled randomly, and a small group will be up-sampled with SMOTE
(synthetic minority over-sampling technique), so that the final group size
will be the total sample number/group number. If it is set as "up", the
samples in a small group will be up-sampled, so its sample size will be
the large group size. If this parameter is set as "dn", the samples in the
large group will be down-sampled so that its sample size will be the small
sample size. Default is "updn".}

\item{nround}{When \code{balanceadj} is TRUE, this parameter will be used
to determine the number of the base learner sets in the ensemble-based
mode. Default is 10.}

\item{seed}{Random seed for the sampling process to generate base learner
sets in ensemble-based limma.}

\item{k}{When \code{balanceadj} is TRUE, and \code{samplingmethod} is "updn"
or "up", this parameter will be used for up-sampling. Because up-sampling
is based on SMOTE, which needs to synthesize new samples from the closet
neighbors of a randomly selected sample in the same group, this parameter
is used to define how many closet neighbors will be needed. Default is 5,
meaning the top 5 closet neighbors will be used for the synthesis.}

\item{threads}{Number of threads need for parallelization. Default is 1.}

\item{method}{When \code{balanceadj} is TRUE, this parameter will be needed
for aggregating the results from all the base learners in ensemble-based
limma. Each base learner will return a set of p-vals for all the features
in the dataset, so for each feature, it will finally get several p-vals
from different base learners, and to combine them into one value, some
p-val combination methods can be used. This parameter is used to choose
the method, can be "Cauchy", "harmonicmean", or "Fisher". Default value is
"Cauchy", which uses ACAT (aggregated Cauchy association test) for the
combination. ACAT is suitable to combine strongly dependent p-vals, which
is the case of ensemble-based limma because its base learner sets are from
the same original dataset.}

\item{removereg}{A GRanges object or a data frame that records the genomic
coordinates of any regions that need to be excluded from the analysis, can
be the genomic regions related to any confounding factors. If a feature
called by normal limma or ensemble-based limma is located there, it will be
removed. For regions related to confounding factors, this region removal
can help avoid influence from the confounding factors. However, because
both normal limma and ensemble-based limma can adjust confoundings during
their regression step, a further region removal process will have very
limited effects. Hence, this step is unnecessary, and can be skipped by
setting this parameter \code{removereg} as NULL. If need to transfer a
data frame, it should have columns named as "seqnames", "start", "end",
and "strand", which help to define the genomic coordinates of the regions.
Each row of the data frame should be a region. The default value is NULL.}

\item{featuretype}{The type of the features in the data. It will be used
when \code{removereg} is not NULL and judge whether any features located
in the genomic regions to be removed. Can be "gene" or "probe" (for probes
in the DNA methylation data). If another parameter \code{plot} is TRUE,
it will also appear in the title of the volcano plot generated by this
function. Default value is "probe".}

\item{platform}{If the data to be analyzed is methylation probe data, this
parameter will be used to indicate the platform of the data. Can be set as
450 (for 450k platform), 850 (for EPIC), or 27 (27K platform). Its default
value is 450.}

\item{ignorestrand}{When the parameter \code{removereg} is not NULL and some
genomic regions need to be excluded from the analysis, this parameter will
be used to judge whether any features located in the genomic regions. If
it is FALSE, the strand information of the features should match that of
the regions if they are judged as within the regions, and if it is TRUE,
the strand information will not be considered. Default is TRUE.}

\item{tssradius}{When \code{removereg} is not NULL and \code{featuretype}
is "gene". This parameter will be used to judge whether a gene located in
any of the genomic regions. If the TSS of a gene is located within the
regions, or not within, but the distance between its TSS and the end of
any regions is less than the value of this \code{tssradius}, this gene
will be judged as located within the regions (if the strand information
required by \code{ignorestrand} is also fulfilled). Then, the gene will
be excluded from the analysis. Default is 1500 (for 1500 bp).}

\item{plot}{Whether need to plot a volcano plot to show the differential
features. Default is FALSE.}

\item{isbetaval}{If the feature values in the data are log2 transformed,
set it as FALSE, if they are not log2 transformed, i.e., they are gene
counts or methylation beta values, set this parameter as TRUE. Will only
be needed when \code{plot} is TRUE.}

\item{absxcutoff}{The cutoff on log2FC (for log2 transformed features) or
inter-group difference (for non-log2 transformed feature counts or beta
values) to judge whether a feature is significantly different between the
sample groups. Default is 0, meaning any log2FC or difference value will
not influence the judgment on feature significance.}

\item{pvalcutoff}{The p-val (or adjusted p-val) cutoff to judge whether a
feature is significantly different between the groups or not. Default is
0.05.}

\item{pvalcolname}{Which p-val should be used to judge the significance of
the features. Can be "P.Val" or "adj.P.Val". Default is "adj.P.Val".}

\item{titleprefix}{The prefix of the volcano plot title. It can be set as
any character string need to be shown in the title. Default is NULL.}

\item{titlesize}{The font size for the plot title of \code{plot}. Default
is 15.}

\item{textsize}{The font size for the plot texts of \code{plot}. Default is
13.}

\item{face}{The font face for the plot texts of \code{plot}. Default value
is "bold".}

\item{labelnum}{In the volcano plot, the names of the top up-regulated and
top down-regulated features can be labeled. This number indicates how many
top features in the groups should be labeled. If it is NULL, no feature
name will be labeled. Can also be any number. Default is NULL.}

\item{annotextsize}{If \code{labelnum} is not NULL, this parameter will be
needed to set the font size of the gene names to be labeled in the volcano
plot. Default is 4.}
}
\value{
If \code{plot} were FALSE, will only return a data frame recording
the normal limma or ensemble-based limma results for all the features in
the data, including the adjusted p-val, log2FC, etc. If \code{plot} were
TRUE, will return a list containing not only the limma results, but also
the plot data, and a volcano plot will be generated.
}
\description{
Perform ensemble-based limma or normal limma to find features relevant to
specific phenotypic variable.
}
