% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_wgcna.R
\name{diffwgcna}
\alias{diffwgcna}
\title{Perform WGCNA analysis}
\usage{
diffwgcna(
  dat,
  pddat,
  responsevarname = NULL,
  responselevels = NULL,
  confoundings = NULL,
  topvaricancetype = "sd",
  topvaricance = 5000,
  balanceadj = FALSE,
  samplingmethod = "updn",
  nround = 10,
  seed = 2022,
  k = 5,
  threads = 1,
  method = "Cauchy",
  sftpowers = NULL,
  powers = seq(1, 20, 1),
  rsqcutline = 0.8,
  minclustersize = 50,
  mergecutheight = 0.2,
  maxblocksize = 5000,
  removereg = NULL,
  featuretype = "gene",
  topoenrichment = FALSE,
  shufflenum = 1000,
  mediation = FALSE,
  topn = NULL,
  platform = 450,
  ignorestrand = TRUE,
  tssradius = 1500,
  plot = FALSE,
  titleprefix = NULL,
  isbetaval = TRUE,
  diffcutoff = 0,
  pvalcutoff = 0.05,
  pvalcolname = "adj.P.Val",
  absxcutoff = 0,
  titlesize = 15,
  textsize = 13,
  face = "bold",
  labelnum = NULL,
  annotextsize = 4,
  savetom = TRUE
)
}
\arguments{
\item{dat}{A matrix with the feature values for samples. Each column is one
sample and each row represents one feature. The row names are the feature
names and the column names should be sample IDs.}

\item{pddat}{Meta data frame. The first column should be the sample names
with a column name as "sampleid", and the remaining should be phenotypic
variables.}

\item{responsevarname}{The column name of the response variable in the data
frame provided to \code{pddat}. It is the variable of interest. Default
is NULL, which means the second column of \code{pddat} is the response.}

\item{responselevels}{If the variable of \code{responsevarname} is discrete,
it will be treated as a factor, and this parameter is needed to define the
factor element level, so it should be a vector defining this level. Only
needed if the response in \code{pddat} is a character vector but not a
factor. In this case, it will be converted to a factor by the function,
and this paramter \code{responselevels} will be used to define the factor
element level. Its default is NULL, meaning that this element level will
be set automatically following the character order of the elements in the
response variable.}

\item{confoundings}{The column name of the confounding factors in the data
frame provided to \code{pddat}. Default is NULL, which means there is no
confounding factor in the data.}

\item{topvaricance}{The number of top variance features need to be included
in the WGCNA analysis. Default is 5000.}

\item{balanceadj}{Whether to use the ensemble-based mode. When it is set
as TRUE, a sampling process will be performed on the samples to generate
several base learner datasets, and the sample groups of each base learner
set will be adjusted during the sampling so that they will have the same
size. After calling the WGCNA modules, a differential feature selection
step will be performed to find the modules with significantly different
eigengene values between the sample groups, and also find the features
with significantly different values between the sample groups and within
each module. At this stage, limma will be used, and if the parameter here
is set as TRUE, limma will be used on each base learner first to call the
differential features. After that, their results will finally be ensembled
together to get the result for the whole data. If this parameter is set as
FALSE, normal limma will be used. Similarly, for the mediation analysis to
find the module features mediating the causal relationships between the
module and the response variable, it can also be performed on the base
learner datasets with an ensemble mode, or directly on the whole data with
a normal mode, depending on this parameter. Default is FALSE.}

\item{samplingmethod}{When \code{balanceadj} is TRUE, this parameter will
be needed to determine how to sample the data to get the base learner sets
for the ensemeble mode. If it is set as "updn", to make the sample groups
have the same sample size in a base leaner dataset, a large group will be
down-sampled randomly, and a small group will be up-sampled with SMOTE
(synthetic minority over-sampling technique), so that the final group size
will be the total sample number/group number. If it is set as "up", the
samples in a small group will be up-sampled, so its sample size will be
the large group size. If this parameter is set as "dn", the samples in the
large group will be down-sampled so that its sample size will be the small
sample size. Default is "updn".}

\item{nround}{When \code{balanceadj} is TRUE, this parameter will be used
to determine the number of the base learner sets in the ensemble-based
mode. Default is 10.}

\item{seed}{Random seed for some random processes, such as base learner sets
generation, shuffling-based functional enrichment, etc.}

\item{k}{When \code{balanceadj} is TRUE, and \code{samplingmethod} is "updn"
or "up", this parameter will be used for up-sampling. Because up-sampling
is based on SMOTE, which needs to synthesize new samples from the closet
neighbors of a randomly selected sample in the same group, this parameter
is used to define how many closet neighbors will be needed. Default is 5,
meaning the top 5 closet neighbors will be used for the synthesis.}

\item{threads}{Number of threads need for parallelization. Default is 1.}

\item{method}{When \code{balanceadj} is TRUE, this parameter will be needed
for aggregating the results from all the base learners in ensemble-based
limma. Each base learner will return a set of p-vals for all the features
in the dataset, so for each feature, it will finally get several p-vals
from different base learners, and to combine them into one value, some
p-val combination methods can be used. This parameter is used to choose
the method, can be "Cauchy", "harmonicmean", or "Fisher". Default value is
"Cauchy", which uses ACAT (aggregated Cauchy association test) for the
combination. ACAT is suitable to combine strongly dependent p-vals, which
is the case of ensemble-based limma because its base learner sets are from
the same original dataset.}

\item{sftpowers}{The soft-thresholding power to be used for WGCNA network
construction. Should be an integer or NULL. Default is NULL, and in this
case, the function will calculate and pick up an appropriate power from
the values provided to another parameter \code{powers}.}

\item{powers}{A vector used to provide candidate soft-thresholding power for
WGCNA network construction. Should be a vector with integers as elements.
If the parameter \code{sftpowers} is NULL, the function will calculate the
scale free topology fitting index for each element of this vector and use
the one with the largest index as the final power. Default is the integer
sequence from 1 to 20.}

\item{rsqcutline}{A float defining the desired minimum scale free topology
fitting index to select the soft-thresholding power from the candidates
included in the parameter \code{powers}, in the case that the parameter
\code{sftpowers} is NULL and the soft-thresholding powers for the data
need to be calculated by the function. Default is 0.8.}

\item{minclustersize}{The minimum module size for module detection during
WGCNA network construction. Default is 50.}

\item{mergecutheight}{The dendrogram cut height for WGCNA module merging. It
default is 0.2.}

\item{maxblocksize}{Integer giving maximum block size for module detection
during WGCNA network construction. Default is 50.}

\item{removereg}{A GRanges object or a data frame that records the genomic
coordinates of any regions that need to be excluded from the analysis, can
be the genomic regions related to any confounding factors. If a feature
is located there, it will be removed at the beginning. For regions related
to confounding factors, this region removal can help avoid influence from
the confounding factors. However, for the limma and mediation analyses,
because they can adjust confoundings, this region removal will have very
limited effects. Hence, this step is unnecessary, and can be skipped by
setting this parameter \code{removereg} as NULL. If need to transfer a
data frame, it should have columns named as "seqnames", "start", "end",
and "strand", which help to define the genomic coordinates of the regions.
Each row of the data frame should be a region. The default value is NULL.}

\item{featuretype}{The type of the features in the data. It will be used
when \code{removereg} is not NULL and judge whether any features located
in the genomic regions to be removed. Can be "gene" or "probe" (for probes
in the DNA methylation data). If another parameter \code{plot} is TRUE,
it will also appear in the title of the dendrogram plot and volcano plot
generated by this function. In addition, only when this \code{featuretype}
is "gene", the functional enrichment analysis will be performed after the
WGCNA network construction. Default value is "gene".}

\item{topoenrichment}{After the WGCNA module detection, this function can
perform network-based gene functional enrichment on each module, i.e.,
the enrichment analysis is performed not only on the gene level, but also
on the gene-gene pair level. If this parameter is TRUE, the enrichment
will be performed. If it is FALSE, this step will be skipped. Default is
FALSE.}

\item{shufflenum}{For the functional enrichment analysis, the function can
perform it with both a hypergeometric enrichment method, and a gene pair
weight or gene weight shuffling enrichment method. This parameter is used
to define the shuffling number for the weight shuffling method. Default is
1000.}

\item{mediation}{After the WGCNA module detection and limma analysis, this
function can perform causal inference (mediation analysis) on each module
with a significant eigengene difference between the sample groups defined
by the parameter \code{responsevarname} in the data frame \code{pddat}. In
each of these modules, each module feature with a significant difference
between the sample groups will be used as the potential mediator of this
mediation analysis (causal inference) and for each feature in each module,
2 kinds of mediation relationships will be tested. The first is "response
(in \code{pddat}) -> module feature -> module eigengene", and the other is
the opposite direction "module eigengene -> module feature -> response".
For the mediation analysis (causal inference), the response variable need
to be either a continuous or a binary variable. If the parameter here is
TRUE, the mediation analysis will be performed to test the above 2 kinds
of mediation relationships and will return the significant ones in the
final result. If this parameter is FALSE, the mediation analysis will not
be performed. Default is FALSE.}

\item{topn}{If the parameter \code{mediation} is set as TRUE, this parameter
\code{topn} will be used to define the number of top differential modules
needed to be analyzed by the mediation analysis. If it is set as 1, the
most differential module between the sample groups will be analyzed by the
mediation model. If it is set as 2, the top 2 most differential modules
will go through the analysis, respectively. If it is set as 3, the top 3
most differential modules will be analyzed, etc. The default value is NULL,
meaning that all the differential modules will go through the mediation
analysis.}

\item{platform}{If the data to be analyzed is methylation probe data, this
parameter will be used to indicate the platform of the data. Can be set as
450 (for 450k platform), 850 (for EPIC), or 27 (27K platform). Its default
value is 450.}

\item{ignorestrand}{When the parameter \code{removereg} is not NULL and some
genomic regions need to be excluded from the analysis, this parameter will
be used to judge whether any features located in the genomic regions. If
it is FALSE, the strand information of the features should match that of
the regions if they are judged as within the regions, and if it is TRUE,
the strand information will not be considered. Default is TRUE.}

\item{tssradius}{When \code{removereg} is not NULL and \code{featuretype}
is "gene". This parameter will be used to judge whether a gene located in
any of the genomic regions. If the TSS of a gene is located within the
regions, or not within, but the distance between its TSS and the end of
any regions is less than the value of this \code{tssradius}, this gene
will be judged as located within the regions (if the strand information
required by \code{ignorestrand} is also fulfilled). Then, the gene will
be excluded from the analysis. Default is 1500 (for 1500 bp).}

\item{plot}{Whether need to generate several plots during the analysis, i.e.
the dendrogram plot and the network plot to show the WGCNA modules, the
bar plot to show the significantly differential WGCNA module eigengenes
between the sample groups, the volcano plots to show the significantly
differential module features between the sample groups and within each
module, the volcano plots to show the significant mediation relationships
within each module. Default is FALSE so that no plots will be generated.}

\item{titleprefix}{The prefix of the titles of the dendrogram plot, network
plot, barplot and volcano plot. It can be set as any character string need
to be shown in the titles. Default is NULL.}

\item{isbetaval}{If the feature values in the data are log2 transformed,
set it as FALSE, if they are not log2 transformed, i.e., they are gene
counts or methylation beta values, set this parameter as TRUE. Will only
be needed when \code{plot} is TRUE.}

\item{diffcutoff}{The cutoff on the inter-group difference to judge whether
the WGCNA module eigengenes are significantly different between the sample
groups in the limma analysis. Default is 0, meaning any difference value
will not influence the judgment of the module eigengene difference, and
the p-val is the only criterion to judge the significance of eigengene
difference.}

\item{pvalcutoff}{The p-val (or adjusted p-val) cutoff to judge whether a
feature is significantly different between the groups or not in the limma
analysis. Default is 0.05.}

\item{pvalcolname}{Which p-val should be used to judge the significance of
the features in the limma analysis. Can be "P.Val" or "adj.P.Val". Default
is "adj.P.Val".}

\item{absxcutoff}{The cutoff on log2FC (for log2 transformed features) or
inter-group difference (for non-log2 transformed feature counts or beta
values) to judge whether a feature is significantly different between the
sample groups in the limma analysis. Default is 0, meaning any log2FC or
difference value will not influence the judgment on feature significance.}

\item{titlesize}{The font size of the barplot and volcano plot titles. Its
default is 15.}

\item{textsize}{The font size of the barplot and volcano plot texts. Default
is 13.}

\item{face}{The font face of the barplot and volcano plot texts. Default is
"bold".
'@param labelnum In the volcano plot, the names of the top up-regulated and
top down-regulated features can be labeled. This number indicates how many
top features in the groups should be labeled. If it is NULL, no feature
name will be labeled. Can also be any number. Default is NULL.}

\item{annotextsize}{If \code{labelnum} is not NULL, this parameter will be
needed to set the font size of the gene names to be labeled in the volcano
plot. Default is 4.}

\item{savetom}{Whether the topological overlap matrix (TOM) of the WGCNA
analysis need to be returned or not. Default is TRUE.}

\item{topvariancetype}{The method need to calculate feature variance. Its
default is 'sd', meaning standard deviation. Can also be 'mad', meaning
median absolute deviation.}
}
\value{
A list with several slots will be returned. One named "wgcnares"
records the WGCNA module detection results, including the module labels
of the features, the soft-thresholding power used for the module calling,
and the module eigengenes of the samples. If the paramter \code{savetom}
is TRUE, this slot will also contain the TOM matrix of the WGCNA analysis.
Another slot named "limmares" records the limma analysis results on the
module eigengene difference between sample groups. The "melimmares" slot
is the limma results on the feature difference between sample groups and
within each module. There are also several slots recording the functional
enrichment results for the modules. The slots named as "genepaths" and
"pairpaths" are the hypergeometric enrichment results on the gene level
and gene-gene pair level, respectively, and the functional terms of the
genes or gene pairs are based on the Reactome database. The gene terms are
directly the ones in this database, and the terms of a gene pair is the
intersection of its 2 genes'. Similarly, the slots named "genebps" and
"pairbps" are the hypergeometric enrichment results on gene and gene pair
levels, but the functional term database is the GOBP database, rather than
Reactome. The result list also has other slots named "geneshufflepaths",
"pairshufflepaths", "geneshufflebps", and "pairshufflebps", which are the
corresponding functional results generated with the gene and gene pair
weight shuffling method, not the hypergeometric test method. It should be
noted that, for the results on gene pairs, because the WGCNA network is
a fully-connected network and the gene-gene combination can generate a
huge number of gene pairs, which will require huge computational cost to
get their functional terms via the intersection computation. Hence, the
hypergeometric results on the gene pairs only restricted to the gene pairs
with the top 1\% weight of the whole WGCNA network. The detailed functional
terms of the gene and gene pairs are in the slots of "geneweightspaths",
"pairweightspaths", and "pairshuffleweightspaths", etc, corresponding to
the results in the slots "genepaths", "pairpaths", and "pairshufflepaths",
etc. For the mediation analysis results, they are included in another slot
named "mediationres", a data frame showing the significant medation casual
relationships for the WGCNA modules. If the mediators are genes, the slots
"mediatorgenedat" contains the biological information of the genes, and if
the mediators are DNA methylation probes, the slot "mediatorprobedat" are
the information of these probes.
}
\description{
Perform WGCNA analysis followed by network-based functional enrichment and
causal inference (based on mediation analysis) for WGCNA modules.
}
