% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_limma.R
\name{featuresampling}
\alias{featuresampling}
\title{Select data features with standard deviation (SD), median absolute deviation
(MAD), or type-III ANOVA.}
\usage{
featuresampling(
  betas,
  topfeatures = 10000,
  pddat = NULL,
  variancetype = "sd",
  anova = FALSE,
  responsevarname = NULL,
  filternovar = TRUE,
  threads = 1,
  plotannovares = TRUE,
  titlesize = 15,
  textsize = 13,
  face = "bold",
  featuretype = NULL,
  plottitlesuffix = NULL
)
}
\arguments{
\item{betas}{Data matrix for feature selection, with features as rows and
samples as columns.}

\item{topfeatures}{The number of features need to select. Default is 10000.}

\item{pddat}{Meta data frame. The first column should be the sample names,
and the remaining should be the phenotypic variables need to be analyzed
with ANOVA. Only needed for ANOVA-based feature selection. Default value
is NULL.}

\item{variancetype}{The method need to calculate feature variance. Default
is 'sd', meaning standard deviation. Can also be 'mad', meaning median
absolute deviation.}

\item{anova}{Whether need to perform ANOVA-based feature selection. Default
is FALSE. If it is TRUE, and \code{pddat} is also provided, the function
will perform ANOVA for each feature in the data, so its variance explained
by each phenotypic variable in \code{pddat} can be obtained. Then, the
features will be ordered by its ANOVA p-val from the response variable in
\code{pddat} (defined by the parameter \code{responsevarname}), and those
with the smallest p-val will be selected and returned. If \code{anova} is
FALSE, or \code{pddat} is not provided, the features with the largest SD
or MAD variance will be selected, as defined by \code{variancetype}.}

\item{responsevarname}{The column name of the response variable in the data
frame provided to \code{pddat}. It is the variable of interest, and other
variables in \code{pddat} will be the confounding variables. Default is
NULL, which means the second column of \code{pddat} will be the response.}

\item{filternovar}{Whether features with no variance across the samples need
to be excluded from the feature selection. Default is TRUE.}

\item{threads}{Number of threads need for parallelization. Default is 1.}

\item{plotannovares}{Whether need to plot the ANOVA result. Default is TRUE.
In this case, the ANOVA results from all the features will be averaged,
including their F statistics, MSS (mean sum of the square), and p-vals,
and the final will be that for the dataset. Then, a bar plot will be made
for this dataset result.}

\item{titlesize}{The font size for the plot title of \code{plotannovares}.
Default is 15.}

\item{textsize}{The font size for the plot texts of \code{plotannovares}.
Default is 13.}

\item{face}{The font face for the plot texts of \code{plotannovares}. The
default value is "bold".}

\item{featuretype}{The type of the features in the data. Will appear in the
title of the ANOVA plot. Can be "gene", "probe", and so on. Default value
is NULL, so that this information will not appear.}

\item{plottitlesuffx}{The prefix of the ANOVA plot title. It can be set as
any character string need to be shown in the title. Default is NULL.}
}
\value{
A list with 2 elements. The first is the final data matrix only with
the selected features. The second is the selection information of all the
features in the original data, such as their SD (if using SD for feature
selection), MAD (if using MAD), or their ANOVA F statistic, MSS, and p-val
(if using ANOVA feature selection). If \code{plotannovares} were TRUE, the
ANOVA result for the dataset will also be plotted.
}
\description{
Select features with the largest SD, largest MAD, or smallest ANOVA p-val.
}
\examples{
library(CWGCNA)

betas <- system.file("extdata", "placentabetas.rds", package = "CWGCNA")
betas <- readRDS(betas)

pds <- system.file("extdata", "placentapds.rds", package = "CWGCNA")
pds <- readRDS(pds)

top10k <- featuresampling(betas = betas, topfeatures = 10000, 
 variancetype = "sd", threads = 6)
 
anovares <- featuresampling(betas = top10k$betas, pddat = pds, anova = TRUE, 
 
 plotannovares = TRUE, featuretype = "probe", plottitlesuffix = "placenta", 
 titlesize = 18, textsize = 16, 
 
 threads = 6)
}
