% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_limma.R
\name{geneanno}
\alias{geneanno}
\title{Annotate genes}
\usage{
geneanno(genesymbols = NULL, geneentrezs = NULL)
}
\arguments{
\item{genesymbols}{A vector containing the symbols of genes to be annotated.
Default is NULL.}

\item{geneentrezs}{A vector with the ENTREZ IDs of genes to be annotated.
If \code{genesymbols} is NULL, this value will be used. Default is NULL.}
}
\value{
A data frame with the gene annotation information.
}
\description{
Annotate gene coordinates and functions
}
