% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_limma.R
\name{imputemeta}
\alias{imputemeta}
\title{Impute missing values in meta data}
\usage{
imputemeta(pddat, percentcutoff = 1/4, seednum = 2022)
}
\arguments{
\item{pddat}{Meta data data frame. The first column should be the sample
names, and the remaining ones should be all the confounding factors in the
data, including both the ones with missing values and the complete ones.}

\item{percentcutoff}{If a confounding factor has a missing value proportion
greater than this \code{percentcutoff} value, it will not be imputed and
will not appear in the final imputed meta data frame returned. Default is
1/4.}

\item{seednum}{A random seed number used by the MICE algorithm to impute the
missing values. Default is 2022.}
}
\value{
A data frame with the original complete confounding factors and the
ones imputed as complete.
}
\description{
Impute missing values in meta data using the MICE (multivariate imputations
by chained equations) method.
}
