% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_mediation.R
\name{lassomediation}
\alias{lassomediation}
\title{Perform multi-omics regulatory analysis}
\usage{
lassomediation(
  methyldat,
  rnadat,
  pddat,
  responsevarname = NULL,
  responselevels = NULL,
  confoundings = NULL,
  calldifffeatures = FALSE,
  topdiffrnafeaturenum = 10,
  balanceadj = TRUE,
  samplingmethod = "updn",
  nround = 10,
  seed = 2022,
  k = 5,
  threads = 1,
  method = "Cauchy",
  methylremovereg = NULL,
  rnaremovereg = NULL,
  ignorestrand = TRUE,
  methylfeaturetype = "probe",
  tssradius = 1500,
  plot = FALSE,
  titleprefix = NULL,
  labelnum = 5,
  isbetaval_methyl = TRUE,
  isbetaval_rna = FALSE,
  absxcutoff = 0,
  pvalcutoff = 0.05,
  pvalcolname = "adj.P.Val",
  titlesize = 15,
  textsize = 13,
  face = "bold",
  annotextsize = 4,
  errortype = "min",
  prescreen = TRUE,
  crossrcut = 0.5,
  methylplatform = 450
)
}
\arguments{
\item{methyldat}{This function performs LASSO plus mediation analysis with
the features in the data \code{rnadat} and \code{methyldat}. The parameter
\code{methyldat} here accepts a matrix with rows as samples and columns as
features, where the row names are sample names and the column names are
feature names. For a feature from \code{rnadat}, the function will use the
LASSO method to select features from \code{methyldat} so that they can fit
the feature from \code{rnadat}. Then, in mediation analysis, that feature
will become the mediator and pair with the LASSO selected \code{methyldat}
features to build mediation models. The data of \code{methyldat} can be a
DNA methylation matrix.}

\item{rnadat}{It is a matrix with each row as a sample and each column as a
feature. The row names are sample names and the column names are feature
names. For a feature from \code{rnadat}, the features in \code{methyldat}
will be used to fit it with LASSO regression. Then, in mediation analysis,
that feature will become the mediator and pair with the LASSO selected
\code{methyldat} features to build mediation models. Typically, the data
of \code{rnadat} is an RNA matrix.}

\item{responsevarname}{Defines which column of \code{pddat} serves as the
response variable in the mediation analysis. It is the column name. If it
is NULL, the second column of \code{pddat} will be the response. For the
mediation analysis, this response variable should be a continuous or a
binary one. Default is NULL.}

\item{responselevels}{If the response variable in the mediation analysis is
binary, it will be converted to a factor by the function. This parameter
is used to define the element level of the factor. It should be a vector
with the elements ordered following the factor level. It can also be NULL,
so that the level will follow the alpha-beta order of the factor elements.
Default is NULL.}

\item{confoundings}{Defines which columns of the \code{pddat} data frame are
the confounding factors in the mediation analysis. It should be a vector
containing the confounding column names. Its default is NULL, meaning no
confoundings need to be included in the mediation analysis.}

\item{calldifffeatures}{The features in \code{methyldat} and \code{rnadat}
can be directly used for the LASSO plus mediation analysis, or they can be
filtered first with limma, and then only the differential features between
sample groups will be used for the LASSO and mediation analysis. If this
parameter is set as TRUE, the differential feature calling step will be
performed, and if it is set as FALSE, it will not be performed. Default is
FALSE.}

\item{topdiffrnafeaturenum}{If \code{calldifffeatures} is set as TRUE, this
parameter will be used to define how many top differential features in the
\code{rnadat} will be used for the LASSO plus mediation analysis. For each
of them, it will be used in the LASSO regression so that the differential
features in \code{methyldat} can fit it. Then, in the mediation analysis,
it will serve as the mediator, and will pair with each \code{methyldat}
feature selected by LASSO. For such a feature pair, 2 mediation directions
will be tested as "response (contained in \code{pddat}) -> RNA feature
(the feature in \code{rnadat}) -> result feature (from \code{methyldat})",
and its opposite as "result feature -> RNA feature -> response". Default
is 10.}

\item{balanceadj}{Whether to use the ensemble-based mode. When it is set
as TRUE, a sampling process will be performed on the samples to generate
several base learner datasets, and the sample groups of each base learner
set will be adjusted during the sampling so that they will have the same
size. Then, for the differential feature selection, limma will be used on
each base learner to call the differential features, and their results
will finally be ensembled together to get the result for the whole data.
If it is set as FALSE, normal limma will be performed. Similarly, for the
LASSO and mediation analysis steps, they can also be performed on the base
learner datasets with an ensemble mode, or directly on the whole data with
a normal mode, depending on this parameter. Default is FALSE.}

\item{samplingmethod}{When \code{balanceadj} is TRUE, this parameter will
be needed to determine how to sample the data to get the base learner sets
for ensemeble mode. If it is set as "updn", to make the sample groups have
the same sample size in a base leaner set, a large group will be randomly
down-sampled randomly, and a small group will be up-sampled with SMOTE
(synthetic minority over-sampling technique), so that the final group size
will be the total sample number/group number. If it is set as "up", the
samples in a small group will be up-sampled, so its sample size will be
the large group size. If this parameter is set as "dn", the samples in the
large group will be down-sampled so that its sample size will be the small
sample size. Default is "updn".}

\item{nround}{When \code{balanceadj} is TRUE, this parameter will be used
to determine the number of the base learner sets in ensemble-based mode.
Default is 10.}

\item{seed}{Random seed for the sampling process to generate base learner
sets in the ensemble-based mode. For the LASSO regression step, it also
defines the 10-fold cross-validation to choose the optimal regularization
constant lambda. The default value is 2022.}

\item{k}{When \code{balanceadj} is TRUE, and \code{samplingmethod} is "updn"
or "up", this parameter will be used for up-sampling. Because up-sampling
is based on SMOTE, which needs to synthesize new samples from the closet
neighbors of a randomly selected sample in the same group, this parameter
is used to define how many closet neighbors will be needed. Default is 5,
meaning the top 5 closet neighbors will be used for the synthesis.}

\item{threads}{Number of threads need for parallelization. Default is 1.}

\item{method}{When \code{balanceadj} is TRUE, this parameter will be needed
for aggregating the results from all the base learners in ensemble-based
limma. Each base learner will return a set of p-vals for all the features
in the dataset, so for each feature, it will finally get several p-vals
from different base learners, and to combine them into one value, some
p-val combination methods can be used. This parameter is used to choose
the method, can be "Cauchy", "harmonicmean", or "Fisher". Default value is
"Cauchy", which uses ACAT (aggregated Cauchy association test) for the
combination. ACAT is suitable to combine strongly dependent p-vals, which
is the case of ensemble-based limma because its base learner sets are from
the same original dataset.}

\item{methylremovereg}{A GRanges object or a data frame with the genomic
coordinates of any regions that need to be excluded from the differential
feature selection step for \code{methyldat}, can be the genomic regions
related to any confounding factors. If a feature called by normal limma or
ensemble-based limma is located there, this feature will be removed. For
regions related to confounding factors, this region removal can help avoid
influence from the confoundings. However, because both normal limma and
ensemble-based limma can adjust confounding factors by themselves, this
further region removal process will have very limited effects. Hence, this
step is unnecessary, and can be skipped by setting this parameter as NULL.
If need to transfer a GRanges or data frame, the data frame should contain
columns named as "seqnames", "start", "end", and "strand", which help to
indicate the genomic coordinates of the regions. Each row of the data frame
should be a region. The default value is NULL.}

\item{rnaremovereg}{Similar to \code{methylremovereg}, but this parameter is
used for the genomic region removal for the limma result of \code{rnadat},
not \code{methyldat}. Default is NULL.}

\item{ignorestrand}{When \code{methylremovereg} or \code{rnaremovereg} is
not NULL and some genomic regions need to be excluded from the analysis,
this parameter will be used to judge whether any features located in the
genomic regions. If it is FALSE, the strand information of the features
should match the regions' strand if they are judged as within the regions,
and if it is TRUE, the strand information will not be considered. Default
is TRUE.}

\item{methylfeaturetype}{The type of the features in the methylation data.
It will be used when \code{methylremovereg} is not NULL and judge whether
any features located in the genomic regions to be removed. Can be "gene"
or "probe" (for probes in the DNA methylation data). If another parameter
\code{plot} is TRUE, it will also appear in the title of the volcano plot
generated by this function. Default value is "probe".}

\item{tssradius}{When \code{rnaremovereg} or \code{methylremovereg} is not
NULL and need to judge whether a gene is in the regions to be removed,
this parameter will be used. If the TSS of a gene is located within the
regions, or not within, but the distance between its TSS and the end of
any regions is less than the value of this \code{tssradius}, this gene
will be judged as located within the regions (if the strand information
required by \code{ignorestrand} is also fulfilled). Then, the gene will
be excluded from the analysis. Default is 1500 (for 1500 bp).}

\item{plot}{Whether need to plot a volcano plot to show the differential
features if the parameter \code{calldifffeatures} is TRUE. It can also
control whether need to plot a heatmap to show the significant mediation
relationships identified.}

\item{titleprefix}{The prefix of the volcano plot title. It can be set as
any character string need to be shown in the title. Default is NULL.}

\item{labelnum}{In the volcano plot, the names of the top up-regulated and
top down-regulated features can be labeled. This number indicates how many
top features in the groups should be labeled. If it is NULL, no feature
name will be labeled. Can also be any number. Default is NULL.}

\item{isbetaval_methyl}{If the feature values in \code{methyldat} are log2
transformed, set it as FALSE, if they are not log2 transformed, i.e., they
are gene counts or methylation beta values, set this parameter as TRUE. It
will only be needed when \code{plot} is TRUE.}

\item{isbetaval_rna}{Similar to \code{isbetaval_methyl}, but it is used to
indicate whether the feature values in \code{rnadat} are log2 transformed
or gene counts, not for \code{methyldat}.}

\item{absxcutoff}{The cutoff on log2FC (for log2 transformed features) or
inter-group difference (for non-log2 transformed feature counts or beta
values) to judge whether a feature is significantly different between the
sample groups. Default is 0, meaning any log2FC or difference value will
not influence the judgment on feature significance, when the parameter
\code{calldifffeatures} is set as TRUE.}

\item{pvalcutoff}{When \code{calldifffeatures} is TRUE, this parameter can
define the p-val (or adjusted p-val) cutoff to judge whether a feature is
significantly different between the groups or not. Default is 0.05.}

\item{pvalcolname}{When \code{calldifffeatures} is TRUE, this parameter can
define which p-val should be used to judge the statistical significance of
the features. Can be "P.Val" or "adj.P.Val". Default is "adj.P.Val".}

\item{titlesize}{The font size of the volcano plot title. Default is 15.}

\item{textsize}{The font size of the volcano plot and heatmap texts. Default
is 13.}

\item{face}{The font face of the volcano plot texts. Default is "bold".}

\item{annotextsize}{If \code{labelnum} is not NULL, this parameter will be
needed to set the font size of the gene names to be labeled in the volcano
plot. Default is 4.}

\item{errortype}{The method to choose the regularization constant lambda for
the LASSO regression. If it is "min", the lambda that gives the minimum
mean 10-fold cross-validation error will be used , and if it is "1ses",
the lambda that gives the cross-validation error within one standard error
of the minimum will be used. The 10-fold cross-validation is defined by
the random seed \code{seed}. The default value of this \code{errortype} is
"min".}

\item{prescreen}{If this parameter is TRUE, before the LASSO training, a
preliminary screen on the features of \code{methyldat} will be performed
by calculating Pearson correlation coefficients to the response variable
(the feature from \code{rnadat}), and the ones with such a p-value less
than 0.05 will be selected for the LASSO model training. Default is TRUE.}

\item{crossrcut}{Before using a feature in \code{rnadat} as the mediator of
mediation analysis, the LASSO model predicted response should have an R
square greater than this \code{crossrcut}, i.e., its Pearson correlation
coefficient with the original \code{rnadat} feature value should be larger
than the square root of this value. Default is 0.5, so the corresponding
Pearson correlation coefficient should be greater than 0.707. If it is not
reached, the mediation analysis for this \code{rnadat} feature will not be
performed.}

\item{methylplatform}{If the data of \code{methyldat} is methylation probe
data, this parameter will be used to indicate the platform of the data.
Can be set as 450 (for 450k platform), 850 (for EPIC platform), or 27 (27K
platform). Its default value is 450.}

\item{pddats}{The phenotypic data of the samples. It is a data frame with
the first column named as "sampleid" and recording the sample names same
as the row names of the matrices \code{rnadat} and \code{methyldat}. It
should also have another column as the response variable in the mediation
analysis, so that two mediation relationships will be tested as "response
(contained in \code{pddat}) -> RNA feature (the feature in \code{rnadat})
-> result feature (from \code{methyldat})", and its opposite as "result
feature -> RNA feature -> response". The column name of the response need
to be transferred to another parameter \code{responsevarname}, so that it
could be identified as the response, but if it is not transferred, the
second column of this \code{pddat} data frame will be the response. For
the mediation analysis, this response variable should be a continuous or a
binary variable.}
}
\value{
A list with several slots will be returned. One named "mediationres"
records the significant mediation relationships from mediation analysis.
Another named "elasticnetcoefs" records the regression coefficients of the
LASSO regression models. The "elasticnetrsquares" slot is the R square of
the LASSO regression models. The "annolist" slot contains the biological
information of the genes or methylation probes with significant mediation
relationships. If the analysis are performed in the ensemble mode, another
slot named "elasticnetweights" will also be returned, which shows the base
learner weights in the ensemble-based LASSO models. In addition, if the
parameter \code{calldifffeatures} is set as TRUE, the differential feature
results will also be returned, with the slots named as "sigg.rnalimma",
"sigg.methyllimma", "nonsigg.rnalimma", and "nonsigg.methyllimma", which
record the significantly and insignificantly differential features in the
data of \code{rnadat} and \code{methyldat}.
}
\description{
Perform multi-omics regulatory analysis between different omics.
}
