% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_enrich.R
\name{locidistribution}
\alias{locidistribution}
\title{Attribute methylation probes to different genomic regions}
\usage{
locidistribution(
  platform = 450,
  allprobes = NULL,
  targetprobelist,
  plotbackground = FALSE,
  removedup = TRUE,
  titlesuffix = NULL,
  titlesize = 20,
  textsize = 15,
  annotextsize = 4,
  face = NULL
)
}
\arguments{
\item{platform}{The platform of the probes need to be annotated. Can be set
as 450 or 850. Default is 450.}

\item{allprobes}{The background probes to perform the Fisher's exact test to
check the enrichment of probes in different genomic regions. Should be a
vector with the names of the background probes as elements. Default value
is NULL, and in this case, all the probes provided by another parameter
\code{targetprobelist} will be used as the background probes together.}

\item{targetprobelist}{A list with each slot recording the probe names of
a specific probe group need to check the genomic region enrichment status
and the names of the slots are the group names. For example, it can be a
list containing 2 slots with one named "Hypermethylated" and the other as
"Hypomethylated", and the probes in each group will be mapped to different
genomic regions and then for each region, its enrichment for the group
probes compared with the background ones will be checked using Fisher's
exact test. For each probe group, a barplot will be generated to show the
probe proportion of each genomic region within that group, and whether the
probes in each genomic region is significantly up or down enriched will be
labeled.}

\item{plotbackground}{If this parameter is TRUE, a barplot showing the probe
proportion of each genomic region in the background will be generated, but
the enrichment significance cannot be labeled. Default is FALSE.}

\item{removedup}{Some methylation probes can be mapped to multiple genomic
island regions or gene regions and for these special ones, whether remove
them from the analysis or not. Defult is TRUE, meaning to remove them.}

\item{titlesuffix}{A suffix string that can be added to the plot title. Its
default value is NULL.}

\item{titlesize}{Font size of the plot title. Default is 20.}

\item{textsize}{Font size of the legend title, legend text, axis label, ect.
Default is 15.}

\item{annotextsize}{Font size of the annotation text in the plot. Default is
4.}

\item{face}{Font face of the plot.}
}
\value{
Will return a list containing 2 slots. One indicates the statistics
after attributing the probes into different island regions, and the other
indicates after attributing them into different gene regions, including
the percentage of the probes in each region, their Fisher's exact p-value,
etc. Also, the corresponding barplots will be generated.
}
\description{
Attribute methylation probes to different genomic regions and show their
distribution patterns
}
