% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_mediation.R
\name{mediationmod}
\alias{mediationmod}
\title{CDE', 'CDE_lower', 'CDE_upper',}
\usage{
mediationmod(
  mediatordat,
  predictdat,
  source,
  target,
  interaction = FALSE,
  bootnum = 100,
  IPW = TRUE,
  responselevels = NULL,
  sourcelevels = NULL,
  threads = 1
)
}
\description{
CDE', 'CDE_lower', 'CDE_upper',
}
