% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_clustering_new.R
\name{multiCCA}
\alias{multiCCA}
\title{Perform PCA/CCA-based omics clustering}
\usage{
multiCCA(
  dats,
  varfeatures = NULL,
  scale = TRUE,
  numcc = 30,
  seednum = 2022,
  perplexity = 4,
  k = c(2),
  consensus = 1,
  threads = 1,
  plot = FALSE,
  titlefix = NULL,
  textsize = 13,
  titlesize = 15,
  face = "bold"
)
}
\arguments{
\item{dats}{A list with each element as each omic data of the samples to be
clustered. Each omic data should be a matrix with columns as samples and
rows as features. The column names are the sample names and the row names
are the feature names.If only one omic data need to be clustered, it need
to be a list with one element as the omic data matrix. If multi-omics data
need to be clustered, the elements (omics) order in the list can influence
the multi-CCA result, and it is suggested to order the elements (omics)
according to their feature numbers, i.e., the omic data matrix with the
largest feature number should be the first element, and the one with the
second largest feature number should be the second element, and so on.}

\item{varfeatures}{For each omic data in \code{dats}, the standard deviation
of the features can be calculated and the top variable features will be
selected and used for clustering. This parameter defines the number of the
top variable features will be selected. It should be an integer, such as
10000, so that the top 10000 most variable features of each omic will be
selected, and if the omic data have a feature number less than 10000, all
the features will be used. Can also be NULL, so that this variable feature
selection step will be skipped and all the omics will use all the features
for the clustering. Default is NULL.}

\item{scale}{Whether the features of the omics need to be standardized, so
that all of them have a mean across all the samples as 0 and a standard
deviation of 1, and the clustering will be performed after that. Default
is TRUE.}

\item{numcc}{If only one omic is provided to \code{dats}, this function will
perform PCA on the data and use the PCs (principle components) with the
largest eigen values for the following k-means clustering to get the final
sample clusters. If mulit-omics data are provided to \code{dats}, CCA will
be used to compress the multi-omics data, and the function will use the
top CCs (canonical components) with the largest singular values for the
following k-means clustering. This parameter \code{numcc} defines how many
top PCs or CCs will be used for the k-means. Default is 30.}

\item{seednum}{Random seed number to be used for the CCA/PCA step fulfilled
with the IRLBA algorithm (implicitly restarted Lanczos bidiagonalization).
It is also used as the random seed for the feature bootstrapping step of
the consensus clustering mode, which can be activated by setting another
parameter \code{consensus} as an integer greater than 1. In addition, it
is also the random seed for the tSNE and UMAP plots to show the clustering
result. Default is 2022.}

\item{perplexity}{The perplexity parameter for the tSNE plot generated to
show the clustering result. Default is 4.}

\item{k}{The CCA/PCA step compresses the multi-omics or single-omic data.
Then, k-means clustring will be used on the compressed data to get the
final clustering result. This parameter is used to transfer the cluster
number to the k-means step. To evaluate the clustering performance, the
function will calculate several internal validation indices, including
Silhouette index, Calinski index, and Davies-Bouldin index. If the optimal
cluster number for the k-means step is not known, can transfer a vector
with the candiate cluster numbers as elements, so k-means clustering will
use each of them as the cluster number and evaluate the clustering results
by calculating the interanl validation indices. The optimal number should
have the largest Silhouette and Calinski indices, but the smallest Davies-
Bouldin index. If the best cluster number is known, it can be transferred
this parameter as a single integer. Then, only it will be used for the
k-means clustering. Default is an integer 2.}

\item{consensus}{If this parameter is set as the integer 1, the normal mode
will be used, i.e., the original data will be used for the CCA/PCA and the
k-means clustering steps. If this parameter is set as an integer > 1, the
consensus mode will be used, and the features in each omic will undergo a
bootstrapping process for several times. Each time, a bootstrapped dataset
will be generated and the CCA/PCA and k-means steps will be used on this
one. This parameter defines the number of bootstrapping. Then, all the
sample clustering results from the bootstrapping data will be aggregated
to get the consensus result. The consensus mode can be very time-consuming
because the CCA/PCA step need to be used on each bootstrapped data. The
default value of this parameter is 1, so that the clustering will be under
the normal mode.}

\item{threads}{Number of threads need for parallelization. Default is 1.}

\item{plot}{Whether need to generate the PCA, tSNE, and UMAP plots to show
the clustering results, and whether to generate the heatmap plots to show
the consensus mode results. Default is FALSE.}

\item{titlefix}{The character string need to be shown in the titles of the
PCA, tSNE, UMAP, and heatmap plots. It can be set as any character string
need to be shown in the titles. Default is NULL.}

\item{textsize}{The font size of the plot texts. Default is 13.}

\item{titlesize}{The font size of the PCA, tSNE, and UMAP plots titles. Its
default is 15.}

\item{face}{The font face of the PCA, tSNE, and UMAP plots texts. Default is
"bold".}
}
\value{
A list with several slots will be returned. The slot named "kreses"
contains the final sample cluster labels for each candidate cluster number
defined by the parameter \code{k}. The slot "ivis" contains the internal
validation indices for each cluster number. The "cca.samples" slot is the
CCA-compressed data. If the consensus clustering mode is used, the slot
"kreses" will be the sample cluster labels for each bootstrapped dataset,
and another slot "consensusres" will be the final consensus cluster labels
for the samples.
}
\description{
Cluster single-omic data with PCA + k-means or cluster multi-omics data with
CCA + k-means.
}
\examples{
library(CWGCNA)

betas <- system.file("extdata", "placentabetas.rds", package = "CWGCNA")
betas <- readRDS(betas)

pds <- system.file("extdata", "placentapds.rds", package = "CWGCNA")
pds <- readRDS(pds)

#Extract the DNAm probe data for the 101 preeclampsia samples
prepds <- subset(pds, Group == "Preeclampsia")
row.names(prepds) <- 1:nrow(prepds)

prebetas <- betas[, prepds$sampleid, drop = FALSE]

#Clustering
presubtyperes <- multiCCA(dats = list(prebetas), 
                         
 k = 2, consensus = 1, 
                         
 seednum = 2022, threads = 6, plot = TRUE, titlefix = "Preeclampsia", 
 titlesize = 18, textsize = 16)



}
