% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_clustering_new.R
\name{pairedensemblepredict}
\alias{pairedensemblepredict}
\title{Predict single-omic or multi-omics sample class}
\usage{
pairedensemblepredict(balanceres, vardats)
}
\arguments{
\item{balanceres}{The model trained by the function \code{omicsclassifier},
which is the slot "mod" of the returned list result of the function.}

\item{dats}{A list with each element as each omic data of the samples to be
classified. Each omic data should be a matrix with columns as features and
rows as samples. The column names are the feature names and the row names
are the sample names. The number of omics here should be the same as the
data used to train the model with the function \code{omicsclassifier}. If
there is only one omic, it need to be a list with one element as the omic
data matrix.}
}
\value{
A list with 2 slots. One named "pres" is a single-column data frame
with the predicted class labels of the samples. The other named "probs" is
a matrix of the predicted class probability distribution of the samples.
}
\description{
Predict single-omic or multi-omics sample class with the model trained from
the function \code{omicsclassifier}.
}
