% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_limma.R
\name{probeanno}
\alias{probeanno}
\title{Annotate Illumina methylation probes}
\usage{
probeanno(platform = 450, probes)
}
\arguments{
\item{platform}{The platform of the probes. Can be 27 (for 27k platform),
450 (for 450k platform), or 850 (for EPIC platform). Default is 450.}

\item{probes}{The names of the probes need to be annotated.}
}
\value{
A data.frame with the probe annotation information, including the
genomic loci of the probes and their related CpG islands and genes.
}
\description{
Annotate Illumina methylation probes
}
