% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_limma.R
\name{probestogenes}
\alias{probestogenes}
\title{Summarize the methylation beta values of probes to genes}
\usage{
probestogenes(
  betadat,
  platform = 450,
  range27k = 200,
  group450k850k = c("TSS200", "TSS1500", "1stExon"),
  includemultimatch = FALSE,
  checkbetaval = TRUE
)
}
\arguments{
\item{betadat}{A matrix recording the beta values of methylation probes for
samples. Each column represents one sample and each row represents one
probe. The row names are the probe names while the column names should be
sample IDs.}

\item{platform}{The platform of the probes. Can be 27 (for 27k platform),
450 (for 450k platform), or 850 (for EPIC platform). Default is 450.}

\item{range27k}{A positive number or a vector with two positive numbers. If
the data is based on 27k platform, it is needed to define which probes
could be considered as related to a specific gene, and only the ones with
a distance to the TSS of a gene less than the maximum value and greater
than the minimum value of \code{range27k} will be considered as related to
the gene, and the beta values of these probes will be averaged to get the
gene beta value. If it is a single number, the probes with a distance less
than this number and greater than 0 will be attributed to a gene. Default
is 200.}

\item{group450k850k}{A vector or single string. If the data is based on 450k
or EPIC platform, this parameter is needed to define which probes could be
considered as related to a specific gene. Only the ones located in the
gene regions included in this parameter will be considered as belong to
the gene. Its value can be selected from "TSS200", "TSS1500", "1stExon",
"5'UTR", '3'UTR", and "Body". Default is the vector c("TSS200", "TSS1500",
"1stExon"), which means probes within these 3 regions of a gene will be
attributed to the gene and their beta values will be averaged to get the
gene beta value.}

\item{includemultimatch}{Several probes can be attributed to more than one
gene. If this parameter is TRUE, these probes will be involved into the
beta value calculation for all their related genes. Otherwise, they will
be discarded, so that the beta values of all the genes are averaged only
from their uniquely related probes. Default is FALSE.}

\item{checkbetaval}{When this parameter is set as TRUE, the function will
check if the data values are between 0 and 1 first, and if not, it will
drop the downstream computation and return NULL. Default is TRUE.}
}
\value{
A matrix recording the summarized gene beta values for samples.
}
\description{
Summarize the methylation beta values of probes to genes by averaging the
probes located closely to the TSS of a gene.
}
\examples{
library(CWGCNA)

betas <- system.file("extdata", "placentabetas.rds", package = "CWGCNA")
betas <- readRDS(betas)

betasgene <- probestogenes(betadat = betas, 
 group450k850k = c("TSS200", "TSS1500", "1stExon"))
}
