% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_limma.R
\name{sigdmrs}
\alias{sigdmrs}
\title{Find DNA methylation regions (DMRs) relevant to specific phenotypic variable}
\usage{
sigdmrs(
  dat,
  pddat,
  responsevarname = NULL,
  responselevels = NULL,
  confoundings = NULL,
  platform = 450,
  maxgap = 300,
  Bvalue = 100,
  randomseed = 2022,
  threads = 1
)
}
\arguments{
\item{dat}{A matrix with the beta values of methylation probes for samples.
Each column represents one sample and each row represents one probe. The
row names are the probe names and the column names should be sample IDs.}

\item{pddat}{Meta data frame. The first column should be the sample names
with a column name as "sampleid", and the remaining should be phenotypic
variables.}

\item{responsevarname}{The column name of the response variable in the data
frame provided to \code{pddat}. It is the variable of interest. Default
is NULL, which means the second column of \code{pddat} is the response.
This response variable should be a binary or continuous variable.}

\item{responselevels}{If the response variable is binary, it will be treated
as a factor, and this parameter is needed to define the factor level, so
it should be a vector defining this level. Only needed if the response in
\code{pddat} is a character vector but not a factor. In this case, it will
be converted to a factor by the function, and \code{responselevels} will
be used to define the factor level. Default is NULL, meaning this level
will be set by the function following the character order of the elements
in the response variable.}

\item{confoundings}{The column name of the confounding factors in the data
frame provided to \code{pddat}. Default is NULL, which means there is no
confounding factor in the data.}

\item{platform}{The platform of the methylation data provided to \code{dat}.
Can be set as 450 (for 450k platform), or 850 (for EPIC platform). Default
is 450.}

\item{maxgap}{An integer indicating the cutoff of probe-probe distance when
clustering the DNA methylation probes in \code{dat} to DMRs. In the DNA
metylation data, if the distance between 2 neighbor probes is less than
\code{maxgap}, they will be clustered into the same DMR. Then, this large
DMR will be further divided into smaller DMRs according to the values of
its probes. For a set of consecutive probes in the large DMR, if all the
probe values were greater than a cutoff value, which is defined by the
function from a bootstrapping distribution, this set will be a small DMR.
On the other hand, if all the probes in a set is less than the minus value
of the cutoff, it will be defined as another small DMR. These small DMRs
will be returned as DMRs significantly related to the response variable.}

\item{Bvalue}{An integer denoting the number of sampling when computing the
bootstrapping distributions. This defaults to 100.}

\item{randomseed}{Random seed for the bootstrapping sampling. Default value
is 2022.}

\item{threads}{Number of threads need for parallelization. Default is 1.}
}
\value{
A data frame recording the DMRs related to the response variable,
with their information such as genomic coordinates, bootstrappoing p-val,
etc.
}
\description{
Cluster the probes into DNA methylation regions (DMRs) and identify the ones
relevant to specific phenotypic variable.
}
