% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_mediation.R
\name{singlelassomediation}
\alias{singlelassomediation}
\title{Perform elastic net regression and/or mediation analysis}
\usage{
singlelassomediation(
  i = 1,
  candrnadats,
  oripddats,
  candmethyldats,
  seednum = 2022,
  threads = 1,
  alpha = 1,
  errortype = "min",
  prescreen = TRUE,
  crossrcut = 0.5,
  plotting = FALSE,
  plotprefix = NULL,
  responsevarname = NULL,
  confoundings = NULL,
  responselevels = NULL,
  methylplatform = 450,
  onlyelasticnet = FALSE,
  onlymediation = FALSE,
  returncompleteres = FALSE,
  propna = TRUE
)
}
\arguments{
\item{i}{This function performs elastic net regression using one feature in
the data \code{candrnadats} as response, or performs mediation test using
it as the potential mediator. The data \code{candrnadats} is a matrix with
rows as samples and columns as features. The parameter \code{i} here is an
column index to select one feature from the multiple features of the data.
Default is 1.}

\item{candrnadats}{The data of the features. It is a matrix with each row as
a sample and each column as a feature. The row names are sample names and
the column names are feature names. The parameter \code{i} will serve as
the column index to select one column from this data as the response of
the elastic net regression or as the potential mediator of the mediation
test. Typically an RNA data matrix.}

\item{oripddats}{The phenotypic data of the samples. It is a data frame with
the first column named as "sampleid" and recording the sample names same
as the row names of the matrix \code{candrnadats} and that of the matrix
\code{candmethyldats}. It should also has another column as the response
variable in the mediation analysis, so that two mediation relationships
will be tested as "response (contained in this \code{oripddats}) -> RNA
feature (the feature in \code{candrnadats} selected by \code{i}) -> result
feature (from \code{candmethyldats})", and its opposite as "result feature
-> RNA feature -> response". The column name of the response variable need
to be transferred to another parameter \code{responsevarname}, so that it
could be identified as the response, but if it is not transferred, the
second column of this \code{oripddats} data frame will be the response.
For the mediation analysis, this response variable should be a continuous
or a binary variable.}

\item{candmethyldats}{The function uses elastic net regression to fit one
feature in \code{candrnadats}, and the predictor features to fit it are
from this parameter \code{candmethyldats}. It is also a matrix with rows
representing samples and columns representing features. All its features
will be used in the elastic net model and then the model will select the
ones able to fit the one feature in \code{candrnadats}. In the mediation
models, the one feature in \code{candrnadats} will serve as the mediator
and it will pair with each feature selected by the elastic net to build
several mediation models. If the elastic net regression step is skipped by
the parameters \code{onlyelasticnet} and \code{onlymediation}, all the
features in \code{candmethyldats} will be used to pair with the one in
\code{candrnadats} to construct several mediation models. The data of this
\code{candmethyldats} is typically a DNA methylation matrix.}

\item{seednum}{The random seed number used for the elastic net regression to
define a 10-fold cross-validation, so an optimal regularization constant
lambda can be selected for the elastic net. Default is 2022.}

\item{threads}{Threads number for parallelization, default is 1.}

\item{alpha}{The alpha parameter for elastic net regression, controlling the
balance of L1 and L2 penalties. Default is 1, meaning the model will only
use L1 penalty, so it is actually a LASSO model.}

\item{errortype}{The method to choose the regularization constant lambda for
the elastic net regression model. If it is set as "min", the lambda that
combines with \code{alpha} and gives the minimum 10-fold cross-validation
error will be used, and if it is set as "1ses", the lambda that combines
with \code{alpha} and gives the cross-validation error within one standard
error of the minimum will be used. The 10-fold cross-validation is defined
by the random seed \code{seednum}. The default value of \code{errortype}
is "min".}

\item{prescreen}{If this parameter is set as TRUE, before the elastic net
training, a preliminary screen on the features of \code{candmethyldats}
will be performed first by calculating Pearson correlation coefficients
to the response variable (the feature from \code{candrnadats}), and the
ones with such a p-value less than 0.05 will be selected for the elastic
net model training. Default is TRUE.}

\item{crossrcut}{If the elastic net regression step is followed by mediation
analysis by the parameters \code{onlyelasticnet} and \code{onlymediation},
the elastic net model predicted response should have an R square greater
than this \code{crossrcut}, i.e., its Pearson correlation coefficient with
the original feature value from \code{candrnadats} should be greater than
the square root of this value. Default value is 0.5, so the corresponding
Pearson correlation coefficient should be greater than 0.707. If it is not
reached, the function will return NULL.}

\item{plotting}{If this parameter is TRUE, a scatter plot will be generated
to compare the elastic net predicted response value and the true response
value (the feature from \code{candrnadats}), and a heatmap will be plotted
to show the elastic net selected features for the fitting (the features in
\code{candmethyldats}). Default is FALSE.}

\item{plotprefix}{The characters that will appear in the title of the plots
when \code{plotting} is TRUE. Default is NULL.}

\item{responsevarname}{Defines which column of \code{oripddats} serves as
the response variable in the mediation analysis. It is the column name.
If it is NULL, the second column of \code{oripddats} will be the response.
For the mediation analysis, this response variable should be a continuous
or a binary one. Default is NULL.}

\item{confoundings}{Defines which columns of \code{oripddats} serve as the
confounding factors in the mediation analysis. It is a vector containing
the confounding column names. Its default is NULL, meaning no confoundings
need to be included in the mediation analysis.}

\item{responselevels}{If the response variable in the mediation analysis is
binary, it will be converted to a factor by the function. This parameter
is used to define the element level of the factor. It should be a vector
with the elements ordered following the factor level. It can also be NULL,
so that the level will follow the alpha-beta order of the factor elements.
Default is NULL.}

\item{methylplatform}{If the matrix transferred to \code{candmethyldats} is
DNA methylation microarray data, this parameter is used to define the data
platform. Can be 27 (for 27k platform), 450 (for 450k platform), or 850
(for EPIC platform). Default is 450.}

\item{onlyelasticnet}{Whether the function need to perform the elastic net
model training only and skip the mediation analysis. Default is FALSE.}

\item{onlymediation}{Whether the function need to perform the mediation test
only and skip the elastic net model training. Default is FALSE.}

\item{returncompleteres}{Whether all the mediation test results need to be
returned, including the insignificant ones. If it is FALSE, only the final
significant results will be returned. Default is FALSE.}

\item{propna}{For the insignificant mediation test relatinships, whether the
insignificant mediation proportion should be returned as it is or directly
labeled as NA because they are insignficant. Default is TRUE, meaning they
will be directly labeled as NA.}
}
\value{
By the parameters \code{onlyelasticnet} and \code{onlymediation}, if
only the elastic net regression is performed, a list will be returned with
slots named as "elasticnetcoefs" and "completecomp". The former records
the regression coefficients of the features from \code{candmethyldats} to
fit the feature in \code{candrnadats}, and the latter is the predicted and
ture values of that feature in \code{candrnadats}. If only the mediation
analysis is performed, a list with a slot named as "mediationres" will be
returned, which records the significant mediation relationships from the
mediation analysis. In addition, other slots named as "probeannores" and
"geneannores" may also included in this result list if the features in the
significant mediation relationships have definite probe and gene function
information. Besides, if the parameter \code{returncompleteres} is set as
TRUE, all the tested mediaiton relationship results will be returned in a
slot named "completeres", no matter they are significant or not. If both
the elastic net regresson and the mediation analysis are performed by the
parameters \code{onlyelasticnet} and \code{onlymediation}, all the slots
above will be returned, including "elasticnetcoefs", "mediaitonres", etc.
However, if no siginficant mediation relationship can be detected by the
mediation analysis step, or the elastic net regression does not pass the
R square cutoff defined by the parameter \code{crossrcut}, the final list
will not have the slots "mediationres", "probeannores", and "geneannores".
}
\description{
Perform elastic net regression and/or mediation analysis between different
omics.
}
