% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_mediation.R
\name{singlemediation}
\alias{singlemediation}
\title{CDE', 'CDE_lower', 'CDE_upper',}
\usage{
singlemediation(
  j,
  mediatordat,
  predictdat,
  target,
  source,
  confoundings,
  interaction = FALSE,
  IPW,
  responselevels,
  mod11form,
  mod1form,
  mod2form
)
}
\description{
CDE', 'CDE_lower', 'CDE_upper',
}
