% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_enrich.R
\name{topoenrich}
\alias{topoenrich}
\title{Perform structure-based gene functional enrichment or normal enrichment}
\usage{
topoenrich(
  geneweights = NULL,
  backweights = NULL,
  genepairweights = NULL,
  backpairweights = NULL,
  genecount = FALSE,
  paircount = FALSE,
  addzeroedgesforshuffle = TRUE,
  threads = 1,
  datasets = c("bp", "cc", "mf", "kegg", "reactome"),
  combinedatasetreses = TRUE,
  seed = 2022,
  shufflenum = 1000,
  shuffletwotail = FALSE
)
}
\arguments{
\item{geneweights}{If need to perform enrichment analysis on the gene level,
this parameter can be used. It accepts a data frame with columns as "GENE"
and "weight". The former records gene symbols or ENTREZ IDs, or mixed gene
names as sybmol::ENTREZ, such as "CACNA1S::779", where "CACNA1S" is the
gene symbol and "779" is the ENTREZ ID. The other column "weight" records
the gene weights, so that a gene weight-based enrichment analysis can be
performed from it. The traditional gene count-based analysis is a special
case of this gene weight-based one, with the weights of all the genes as
\enumerate{
\item This parameter can also accept a vector of gene symbols or ENTREZ IDs,
or the mixed gene names, and all the genes in this vector will be assigned
a weight as 1, so the gene count-based analysis will be performed. Default
is NULL, which means the enrichment will not be performed on gene level,
because this function can also perform the enrichment on gene-gene pair
level.
}}

\item{backweights}{The background for the gene level enrichment. Similar to
\code{geneweights}, it can be a data frame with columns named as "GENE"
and "weight", or a vector recording the gene names. The difference is that
it is used to transfer the gene background to perform the hypergeometric
enrichment, not to transfer the genes to be analyzed. Default is NULL, and
in this case, all the human genomic genes will be used as the background.
In this case, the weights of all the genes in the background will be 1 and
the enrichment will be the gene count-based one, not the weight-based one.
The background is only needed for the hypergeometric enrichment analysis.
The function can also perform a gene weight shuffling enrichment, and in
this case, the background is not needed.}

\item{genepairweights}{If need to perform enrichment analysis on the gene-
gene pair level, this parameter can be used. It accepts a data frame with
columns as "fromNode", "toNode" and "weight". The "fromNode" and "toNode"
columns are similar to the "GENE" column in \code{geneweights}, recording
gene symbols or ENTREZ IDs, or mixed gene names. Another column "weight"
records the gene pair weights, so that a gene pair weight-based enrichment
analysis can be performed from it. It can also perform a gene pair count-
based enrichment if all the values in the column "weight" are 1, or there
is no column named "weight" in the data frame. The gene pair count-based
enrichment is a special case of the weight-based one. The default value of
this parameter is NULL, which means the enrichment will not be performed
on gene-gene pair level, because this function can also perform the a gene
level enrichment analysis.}

\item{backpairweights}{The background for the gene pair level enrichment.
Similar to \code{genepairweights}, it should be a data frame with columns
named as "fromNode", "toNode" and "weight". The difference is that it is
used to transfer the gene pair background to perform the hypergeometric
enrichment, not to transfer the gene pairs to be analyzed. Its default is
NULL, and in this case, the gene pair weight-based hypergeometric analysis
will not be performed, because this gene pair background is only necessary
for hypergeometric enrichment, and the function will also perform a gene
pair weight shuffling enrichment, where the background is not needed.}

\item{genecount}{For the enrichment on gene level, if this parameter is set
as TRUE, all the gene weights will be set as 1 and the enrichment will be
performed on the gene count level. Default is FALSE.}

\item{paircount}{For the enrichment on gene pair level, if this parameter is
set as TRUE, all the pair weights will be set as 1 and the enrichment will
be performed on the gene pair count level. Default is FALSE.}

\item{addzeroedgesforshuffle}{For the gene pair weight shuffling enrichment,
if 2 genes do not constitute a pair, e.g., in a gene network, there is no
edge between 2 genes, but this parameter is set as TRUE, the gene pair can
still participate in the gene pair weight shuffling process, just with a
weight of 0. If it is set as FALSE, this gene pair will not participate in
the shuffling process. Default is TRUE.}

\item{threads}{Number of threads to be used for parallelization, default is
1.}

\item{combinedatasetreses}{If transferring multiple functional datasets to
the parameter \code{datasets}, and this parameter is set as FALSE, each
dataset will get a separate results set for the enrichment from it. On the
other hand, if this parameter is TRUE, all the results will be combined
together and then returned. Default is TRUE.}

\item{seed}{Random seed for the gene or gene pair weight shuffling method.
Default is 2022.}

\item{shufflenum}{The shuffling number for the weight shuffling method. Its
default value is 1000.}

\item{shuffletwotail}{Whether the enrichment p-val returned by the shuffling
method should be two tailed or single tailed. Default is FALSE, meaning it
should be a single tailed one, representing the probability of obtaining a
greater weight sum for the genes or pairs belonging to a functional term.}

\item{datasests}{The gene functional datasets to be used for the gene and
gene pair level enrichment, can be a vector including any of the 5 dataset
names: "bp", "cc", "mf", "kegg", and "reactome", representing the GOBP,
GOCC, GOMF, KEGG, and REACTOME datasets. Default is with all these 5.}
}
\value{
A list with several slots. If gene level enrichment is performed, it
will contain a slot named "geneenrichres", which includes the gene level
hypergeometric enrichment, weight shuffling enrichment results, and the
functional terms of each gene. If gene pair level enrichment is performed,
it will contain a slot named "pairenrichres", which includes the gene pair
level hypergeometric enrichment, weight shuffling enrichment results, and
the functional terms of each gene pair, which are the intersection terms
of its 2 genes' functional terms. These results can also be the results
based on gene counts and pair counts, instead of weights. If the parameter
\code{combinedatasetreses} is FALSE, the results can be further split into
different dataset results according to the ones assigned by the parameter
\code{datasets}.
}
\description{
Perform structure-based gene functional enrichment or normal enrichment.
}
