% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emethyl_clustering_new.R
\name{wgcnacluster}
\alias{wgcnacluster}
\title{Perform WGCNA-based omics clustering}
\usage{
wgcnacluster(
  dats,
  varfeatures = 5000,
  scale = TRUE,
  sftpowers = NULL,
  powers = seq(1, 20, 1),
  rsqcutline = 0.8,
  minclustersize = 50,
  mergecutheight = 0.2,
  maxblocksize = 5000,
  seednum = 2022,
  perplexity = 4,
  k = c(2),
  consensus = 1,
  threads = 1,
  plot = TRUE,
  titlefix = NULL,
  textsize = 13,
  titlesize = 15,
  face = "bold"
)
}
\arguments{
\item{dats}{A list with each element as each omic data of the samples to be
clustered. Each omic data should be a matrix with columns as samples and
rows as features. The column names are the sample names and the row names
are the feature names.If only one omic data need to be clustered, it need
to be a list with one element as the omic data matrix.}

\item{varfeatures}{For each omic data in \code{dats}, the standard deviation
of the features can be calculated and the top variable features will be
selected and used for clustering. This parameter defines the number of the
top variable features will be selected. It should be an integer, such as
10000, so that the top 10000 most variable features of each omic will be
selected, and if the omic data have a feature number less than 10000, all
the features will be used. Can also be NULL, so that this variable feature
selection step will be skipped and all the omics will use all the features
for the clustering. Default is 5000, so that for each omic, the top 5000
most variable features will be used for WGCNA module calling, and then,
the WGCNA module eigengenes from all the omics will be combined together
into one compressed data followed by k-means clustering.}

\item{scale}{Whether the features of the omics need to be standardized, so
that all of them have a mean across all the samples as 0 and a standard
deviation of 1, and the clustering will be performed after that. Default
is TRUE.}

\item{sftpowers}{The soft-thresholding power to be used for WGCNA network
construction. Should be an integer or NULL. Default is NULL, and in this
case, for each omic, the function will respectively calculate and pick up
an appropriate power for it, from the values provided to another parameter
\code{powers}. If \code{sftpowers} is an integer, all the omic data will
use this same soft-thresholding power for the WGCNA network construction.}

\item{powers}{A vector used to provide candidate soft-thresholding power for
WGCNA network construction. Should be a vector with integers as elements.
If the parameter \code{sftpowers} is NULL, for each omic, the function
will calculate the scale free topology fitting index for each element of
this vector and use the one with the largest index as the final power. Its
default value is the integer sequence from 1 to 20.}

\item{rsqcutline}{A float defining the desired minimum scale free topology
fitting index to select the soft-thresholding power from the candidates
included in the parameter \code{powers}, in the case that the parameter
\code{sftpowers} is NULL and the soft-thresholding powers for the omics
need to be calculated by the function. Default is 0.8.}

\item{minclustersize}{The minimum module size for module detection during
WGCNA network construction. Default is 50.}

\item{mergecutheight}{The dendrogram cut height for WGCNA module merging. It
default is 0.2.}

\item{maxblocksize}{Integer giving maximum block size for module detection
during WGCNA network construction. Default is 50.}

\item{seednum}{Random seed number to be used for the WGCNA module calling
step. It is also used as the random seed for the feature bootstrapping
step of the consensus clustering mode, which can be activated by setting
another parameter \code{consensus} as an integer greater than 1. It is
also the random seed for the tSNE and UMAP plots to show the clustering
result. Default is 2022.}

\item{perplexity}{The perplexity parameter for the tSNE plot generated to
show the clustering result. Default is 4.}

\item{k}{The WGCNA step compresses the multi-omics or single-omic data by
combining the omics WGCNA eigengenes. Then, k-means clustring will be used
on the compressed data to get the final clustering result. This parameter
is used to transfer the cluster number to the k-means step. To evaluate
the clustering performance, the function will calculate several internal
validation indices, including Silhouette index, Calinski-Harabasz index,
and Davies-Bouldin index. If the optimal cluster number for the k-means
step is not known, can transfer a vector with the candiate cluster numbers
as elements, so k-means clustering will use each of them as the cluster
number and evaluate their clustering results by calculating the interanl
validation indices. The optimal number should have the largest Silhouette
and Calinski-Harabasz indices, but the smallest Davies-Bouldin index. If
the best cluster number is known, it can be transferred to the parameter a
single integer so only it will be used for the k-means clustering. Default
is an integer 2.}

\item{consensus}{If this parameter is set as the integer 1, the normal mode
will be used, i.e., the original data will be used for the WGCNA and the
k-means clustering steps. If this parameter is set as an integer > 1, the
consensus mode will be used, and the features in each omic will undergo a
bootstrapping process for several times. Each time, a bootstrapped dataset
will be generated and the WGCNA and k-means steps will be used on this
one. This parameter defines the number of bootstrapping. Then, all the
sample clustering results from the bootstrapping data will be aggregated
to get the consensus result. The consensus mode can be very time-consuming
because the WGCNA step need to be performed on each bootstrapped data. The
default value of this parameter is 1, so that the clustering will be under
the normal mode.}

\item{threads}{Number of threads need for parallelization. Default is 1.}

\item{plot}{Whether need to generate the WGCNA dendrogram plot during the
WGCNA module calling, whether to generate the PCA, tSNE, and UMAP plots to
show the clustering results, and whether to generate the heatmap plots to
show the consensus mode results. Default is FALSE.}

\item{titlefix}{The character string need to be shown in the titles of the
WGCNA dendrogram, PCA, tSNE, UMAP, and heatmap plots. It can be set as any
character string need to be shown in the titles. Default is NULL.}

\item{textsize}{The font size of the plot texts. Default is 13.}

\item{titlesize}{The font size of the PCA, tSNE, and UMAP plots titles. Its
default is 15.}

\item{face}{The font face of the PCA, tSNE, and UMAP plots texts. Default is
"bold".}
}
\value{
A list with several slots will be returned. The slot named "kreses"
contains the final sample cluster labels for each candidate cluster number
defined by the parameter \code{k}. The slot "ivis" contains the internal
validation indices for each cluster number. The slot "wgcna.samples" has
the WGCNA-compressed data. If the consensus clustering mode is used, the
slot "kreses" will contain the sample cluster labels for each bootstrapped
dataset, and another slot "consensusres" will contain the final consensus
cluster labels for the samples.
}
\description{
Cluster single-omic or multi-omics data with WGCNA + k-means.
}
